+++ packages/components/components/elvis-progressbar/src/web_component/elvis-progressbar.ts
@@ -1,0 +1,15 @@
+import * as ReactProgressbarrComponent from '../../../react.js';
+import { ElviaComponent } from '@elvia/component';
+const style = `{{INSERT_STYLE_HERE}}`;
+
+export default class ElvisProgressbar extends ElviaComponent {
+  static get observedAttributes(): any[] {
+    return [];
+  }
+
+  constructor() {
+    super(ElvisProgressbar, ReactProgressbarrComponent.Progressbar, style, '');
+  }
+}
+
+window.customElements.define('elvis-progressbar', ElvisProgressbar as any);