+++ packages/components/elvia-components.config.js
@@ -1,5 +1,6 @@
 module.exports = [
     { name: 'elvia-checkbox' },
     { name: 'elvia-popover' },
+    { name: 'elvis-progressbar' },
     { name: 'elvia-component' },
 ]